# rlp: pool decode streams