This pr introduces a pool of Stream objects to eliminate the need to reallocate memory for every new Stream. It should help to improve the performance of every call to Decode and Decode bytes. 

New streams will be allocated whenever needed.