@MariusVanDerWijden 2019-02-13
Yes I know, I've rewritten it, will update shortly