@MariusVanDerWijden 2019-02-14
@fjl can you review again? 
It significantly drops the time and amount of memory allocated for repeated calls to web3.getBlock()