@MariusVanDerWijden 2019-02-19
Anyone? @fjl @holiman @karalabe 
It's the same pattern used in log, swarm and trie

 https://github.com/ethereum/go-ethereum/blob/a9835c1816bc49ee54c82b4f2a5b05cbcd89881b/log/format.go#L360